pyproject.toml
[project]
name = "caviardeur"
version = "0.1.0"
description = "Pseudonymize PII in documents locally (PDF, DOCX, PPTX, XLSX, TXT, MD, JSON, XML)"
readme = "README.md"
authors = [
    { name = "rabadin" }
]
requires-python = ">=3.12,<3.15"
dependencies = [
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "pymupdf>=1.24",
    "transformers>=4.40,<5",
    "torch>=2.2",
    "sentencepiece>=0.2",
    "protobuf>=4.0",
    "python-magic>=0.4",
    "click>=8.1",
    "rich>=13.0",
]

[project.scripts]
caviardeur = "caviardeur.cli:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pyinstaller>=6.0",
    "ruff>=0.8",
    "ty>=0.0.17",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=caviardeur --cov-report=term-missing --cov-fail-under=80"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
